Write functionalitites of download.py
Wrote functionalities of accepting dates from a fzf list and update it
if needed. Code to convert date from list to sunpy valid date and append
user input time to it. Selecting wavelength, but can be improved
probably as shell=True is not recommended for some reason. Then
downloading data depending on sampling time and instrument.